screen: move functions to niimpy.screen
